requirements.txt
typer[all]
gitpython
openai
requests
keyring
tqdm
python-dotenv
sentence-transformers
faiss-cpu
google-generativeai
anthropic